% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capesData.R
\docType{data}
\name{capesData}
\alias{capesData}
\title{Data on Scholarships in CAPES International Mobility Programs}
\format{
A data frame with 146036 rows and 14 columns:
\describe{
\item{\strong{ID_PROCESSO}}{Number of concession process.}
\item{\strong{NM_NIVEL}}{Type of scholarship.}
\item{\strong{AN_INICIO}}{Year in which the scholarship begins.}
\item{\strong{ME_INICIO}}{Month in which the scholarship begins.}
\item{\strong{AN_FIM}}{End year of the scholarship.}
\item{\strong{ME_FIM}}{End month of the scholarship.}
\item{\strong{NM_IES_ORIGEM_PRINCIPAL_DA}}{Name of the beneficiary's home institution.}
\item{\strong{NM_UF_IES_ORIGEM}}{Name of the Federation State of the beneficiary's institution of origin.}
\item{\strong{NM_IES_ESTUDO_PRINCIPAL_DA}}{Name of the beneficiary's study institution.}
\item{\strong{NM_PAIS_IES_ESTUDO}}{Country of destination of the beneficiary.}
\item{\strong{NM_GRANDE_AREA}}{Field of knowledge.}
\item{\strong{NM_AREA_CONHECIMENTO}}{Area of knowledge classified by CAPES.}
\item{\strong{NM_AREA_AVALIACAO}}{Area of project evaluation.}
\item{\strong{NM_PROGRAMA}}{Program Name.}
}
}
\source{
Created by Leonardo Biazoli with data provided by CAPES/BRAZIL \url{https://dadosabertos.capes.gov.br/dataset/}.
}
\usage{
capesData
}
\description{
The goal of capesData is to provide an attractive dataset for exploring and visualizing data on the International Mobility Programs funded by CAPES/Brazil. The CAPES database refers to international mobility programs for the period from 2010 to 2019.
}
\examples{
capesData
}
\seealso{
\link{capesData_raw}
}
\keyword{datasets}
